requirements.txt
# AI Code Review Assistant Dependencies

# Terminal UI
rich>=13.7.0

# AI Model Integration
ollama>=0.1.7

# Git Operations
gitpython>=3.1.40

# CLI Prompts
prompt-toolkit>=3.0.43

